+++ libraries/PCF8574/PCF8574.h
@@ -3,7 +3,7 @@
 //    FILE: PCF8574.h
 //  AUTHOR: Rob Tillaart
 //    DATE: 02-febr-2013
-// VERSION: 0.3.9
+// VERSION: 0.4.0
 // PURPOSE: Arduino library for PCF8574 - 8 channel I2C IO expander
 //     URL: https://github.com/RobTillaart/PCF8574
 //          http://forum.arduino.cc/index.php?topic=184800
@@ -13,7 +13,7 @@
 #include "Wire.h"
 
 
-#define PCF8574_LIB_VERSION         (F("0.3.9"))
+#define PCF8574_LIB_VERSION         (F("0.4.0"))
 
 #ifndef PCF8574_INITIAL_VALUE
 #define PCF8574_INITIAL_VALUE       0xFF
@@ -29,9 +29,6 @@ class PCF8574
 public:
   explicit PCF8574(const uint8_t deviceAddress = 0x20, TwoWire *wire = &Wire);
 
-#if defined (ESP8266) || defined(ESP32)
-  bool    begin(int sda, int scl, uint8_t value = PCF8574_INITIAL_VALUE);
-#endif
   bool    begin(uint8_t value = PCF8574_INITIAL_VALUE);
   bool    isConnected();