+++ libraries/PCF8574/test/unit_test_001.cpp
@@ -55,6 +55,7 @@ unittest(test_begin)
 {
   PCF8574 PCF(0x38);
 
+  Wire.begin();
   PCF.begin();
 
   int readValue = PCF.read8();
@@ -70,6 +71,7 @@ unittest(test_read)
   PCF8574 PCF(0x38);
   int readValue;
 
+  Wire.begin();
   PCF.begin();
   for (int i = 0; i < 8; i++)
   {
@@ -92,7 +94,9 @@ unittest(test_address)
 {
   PCF8574 PCF(0x38);
 
-  // incorrect in test environment.
+  Wire.begin();
+
+  //  incorrect in test environment.
   assertTrue(PCF.begin());
   assertTrue(PCF.isConnected());
   assertEqual(0x38, PCF.getAddress());